+++ discovery-provider/src/api/v1/helpers.py
@@ -31,51 +31,19 @@ redis = get_redis()
 logger = logging.getLogger(__name__)
 
 
-PROFILE_PICTURE_SIZES = ["150x150", "480x480", "1000x1000"]
-PROFILE_COVER_PHOTO_SIZES = ["640x", "2000x"]
-COVER_ART_SIZES = ["150x150", "480x480", "1000x1000"]
-
-
 def make_image(endpoint, cid, width="", height=""):
     return f"{endpoint}/content/{cid}/{width}x{height}.jpg"
-
-
-def init_rendezvous(user, cid):
-    if not cid:
-        return ""
-    healthy_nodes = get_all_healthy_content_nodes_cached(redis)
-    if not healthy_nodes:
-        logger.error(
-            f"No healthy Content Nodes found for fetching cid for {user.user_id}: {cid}"
-        )
-        return ""
-
-    return RendezvousHash(
-        *[re.sub("/$", "", node["endpoint"].lower()) for node in healthy_nodes]
-    )
-
-
-def get_primary_endpoint(user, cid):
-    rendezvous = init_rendezvous(user, cid)
-    if not rendezvous:
-        return ""
-    return rendezvous.get(cid)
-
-
-def get_n_primary_endpoints(user, cid, n):
-    rendezvous = init_rendezvous(user, cid)
-    if not rendezvous:
-        return ""
-    return rendezvous.get_n(n, cid)
 
 
 def add_track_artwork(track):
     if "user" not in track:
         return track
     cid = track["cover_art_sizes"]
-    cover_cids = get_image_cids(track["user"], cid, COVER_ART_SIZES)
+    cover_cids = api_helpers.get_image_cids(
+        track["user"], cid, api_helpers.COVER_ART_SIZES
+    )
     track["cover_art_cids"] = cover_cids
-    endpoint = get_primary_endpoint(track["user"], cid)
+    endpoint = api_helpers.get_primary_endpoint(track["user"], cid)
     artwork = get_image_urls(track["user"], cover_cids)
     if endpoint and not artwork:
         # Fallback to legacy image url format with dumb endpoint
@@ -93,9 +61,11 @@ def add_playlist_artwork(playlist):
         return playlist
 
     cid = playlist["playlist_image_sizes_multihash"]
-    cover_cids = get_image_cids(playlist["user"], cid, COVER_ART_SIZES)
+    cover_cids = api_helpers.get_image_cids(
+        playlist["user"], cid, api_helpers.COVER_ART_SIZES
+    )
     playlist["cover_art_cids"] = cover_cids
-    endpoint = get_primary_endpoint(playlist["user"], cid)
+    endpoint = api_helpers.get_primary_endpoint(playlist["user"], cid)
     artwork = get_image_urls(playlist["user"], cover_cids)
     if endpoint and not artwork:
         # Fallback to legacy image url format with dumb endpoint
@@ -129,9 +99,11 @@ def add_user_artwork(user):
     user["profile_picture_legacy"] = user.get("profile_picture")
 
     profile_cid = user.get("profile_picture_sizes")
-    profile_endpoint = get_primary_endpoint(user, profile_cid)
+    profile_endpoint = api_helpers.get_primary_endpoint(user, profile_cid)
     if profile_cid:
-        profile_cids = get_image_cids(user, profile_cid, PROFILE_PICTURE_SIZES)
+        profile_cids = api_helpers.get_image_cids(
+            user, profile_cid, api_helpers.PROFILE_PICTURE_SIZES
+        )
         user["profile_picture_cids"] = profile_cids
         profile = get_image_urls(user, profile_cids)
         if profile_endpoint and not profile:
@@ -143,9 +115,11 @@ def add_user_artwork(user):
             }
         user["profile_picture"] = profile
     cover_cid = user.get("cover_photo_sizes")
-    cover_endpoint = get_primary_endpoint(user, cover_cid)
+    cover_endpoint = api_helpers.get_primary_endpoint(user, cover_cid)
     if cover_cid:
-        cover_cids = get_image_cids(user, cover_cid, PROFILE_COVER_PHOTO_SIZES)
+        cover_cids = api_helpers.get_image_cids(
+            user, cover_cid, api_helpers.PROFILE_COVER_PHOTO_SIZES
+        )
         user["cover_photo_cids"] = cover_cids
         cover = get_image_urls(user, cover_cids)
         if cover_endpoint and not cover:
@@ -162,73 +136,6 @@ def add_user_artwork(user):
 # Helpers
 
 
-async def fetch_url(url):
-    loop = asyncio.get_event_loop()
-    future = loop.run_in_executor(None, requests.get, url)
-    response = await future
-    return response
-
-
-async def race_requests(urls, timeout):
-    tasks = [asyncio.create_task(fetch_url(url)) for url in urls]
-    done, pending = await asyncio.wait(
-        tasks, return_when=asyncio.ALL_COMPLETED, timeout=timeout
-    )
-    for task in done:
-        response = task.result()
-        if response.status_code == 200:
-            return response
-    raise Exception(f"No 200 responses for urls {urls}")
-
-
-# Get cids corresponding to each transcoded variant for the given upload_id.
-# Cache upload_id -> cids mappings.
-def get_image_cids(user, upload_id, variants):
-    if not upload_id:
-        return {}
-    try:
-        image_cids = {}
-        if upload_id.startswith("Qm"):
-            # Legacy path - no need to query content nodes for image variant cids
-            image_cids = {variant: f"{upload_id}/{variant}.jpg" for variant in variants}
-        else:
-            redis_key = f"image_cids:{upload_id}"
-            image_cids = redis.hgetall(redis_key)
-            if image_cids:
-                image_cids = {
-                    variant.decode("utf-8"): cid.decode("utf-8")
-                    for variant, cid in image_cids.items()
-                }
-            else:
-                # Query content for the transcoded cids corresponding to this upload id and
-                # cache upload_id -> { variant: cid, ... }
-                endpoints = get_n_primary_endpoints(user, upload_id, 3)
-                urls = list(
-                    map(lambda endpoint: f"{endpoint}/uploads/{upload_id}", endpoints)
-                )
-
-                # Race requests in a new event loop
-                new_loop = asyncio.new_event_loop()
-                asyncio.set_event_loop(new_loop)
-                resp = new_loop.run_until_complete(race_requests(urls, 1))
-                new_loop.close()
-
-                resp.raise_for_status()
-                image_cids = resp.json().get("results", {})
-                if not image_cids:
-                    return image_cids
-
-                image_cids = {
-                    variant.strip(".jpg"): cid for variant, cid in image_cids.items()
-                }
-                redis.hset(redis_key, mapping=image_cids)
-                redis.expire(redis_key, 86400)  # 24 hour ttl
-        return image_cids
-    except Exception as e:
-        logger.error(f"Exception fetching image cids for id: {upload_id}: {e}")
-        return {}
-
-
 # Map each image cid to a url with its preferred node. This reduces
 # redirects from initially attempting to query the wrong node.
 def get_image_urls(user, image_cids):
@@ -239,7 +146,7 @@ def get_image_urls(user, image_cids):
     for variant, cid in image_cids.items():
         if variant == "original":
             continue
-        primary_endpoint = get_primary_endpoint(user, cid)
+        primary_endpoint = api_helpers.get_primary_endpoint(user, cid)
         if not primary_endpoint:
             raise Exception(
                 "Could not get primary endpoint for user {user.user_id}, cid {cid}"
